backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
python-multipart
emails
email-validator
